feat: move some logic into utils file
