+++ validate/string/src/index.ts
@@ -1,1 +1,2 @@
 export * from './validate.string.plugin'
+export * from './validate.string.utils'
